requirements.txt
python-telegram-bot
matplotlib
reportlab